Separate login into random and login utils
