+++ modules/error.js
@@ -3,9 +3,9 @@
 // Dependencies
 const path = require("path"); // URI and local file paths
 
-// Custom Modules
-const customModulePath = __dirname;
-const login = require(path.join(customModulePath, "login.js"));
+// Utility Modules
+const utilityModulesPath = path.join(__dirname, "utilityModules");
+const loginUtils = require(path.join(utilityModulesPath, "loginUtils.js"));
 
 const accessDeniedStatusCode = 403;
 const notFoundStatusCode = 404;
@@ -58,7 +58,7 @@ exports.handleUnexpectedError = async function(err, req, res, next)
 // Helper Functions
 async function getErrorPageData(req, res)
 {
-    const isUserLoggedIn = await login.isUserLoggedIn(req, res);
+    const isUserLoggedIn = await loginUtils.isUserLoggedIn(req, res);
     const errorPageData = {
         isLoginError: !isUserLoggedIn
     };
