+++ modules/landing.js
@@ -5,12 +5,12 @@ const path = require("path"); // URI and local file paths
 
 // Custom Modules
 const customModulePath = __dirname;
-const login = require(path.join(customModulePath, "login.js"));
 const home = require(path.join(customModulePath, "home.js"));
 
 // Utility Modules
 const utilityModulesPath = path.join(__dirname, "utilityModules");
 const logger = require(path.join(utilityModulesPath, "logger.js"));
+const loginUtils = require(path.join(utilityModulesPath, "loginUtils.js"));
 
 // Landing Logic
 exports.getLandingPage = async function(req, res, next)
@@ -18,7 +18,7 @@ exports.getLandingPage = async function(req, res, next)
     try
     {
         // Try to get the home page if the user is already logged in or can authenticate
-        const isUserLoggedIn = await login.isUserLoggedIn(req, res);
+        const isUserLoggedIn = await loginUtils.isUserLoggedIn(req, res);
         if (!isUserLoggedIn)
         {
             throw new Error("User is not logged in. Falling back to landing page.");
