+++ modules/utilityModules/encoding.js
@@ -1,3 +1,12 @@
+"use strict";
+
+// Dependencies
+const path = require("path"); // URI and local file paths
+
+// Utility Modules
+const utilityModulesPath = __dirname;
+const logger = require(path.join(utilityModulesPath, "logger.js"));
+
 // Encoding Logic
 exports.encodeInBase64 = function(stringToEncode)
 {
@@ -12,4 +21,4 @@ exports.encodeInBase64 = function(stringToEncode)
         logger.logError(`Failed to base 64 encode string: ${stringToEncode}`);
         return Promise.reject(error);
     }
-}
+};
