+++ modules/utilityModules/random.js
@@ -1,0 +1,14 @@
+"use strict";
+
+// Random Logic
+exports.generateRandomString = function(targetLength)
+{
+    let text = "";
+    const possible = "ABCDEFGHIJKLMNOPQRSTUVWXYZabcdefghijklmnopqrstuvwxyz0123456789";
+
+    for (let i = 0; i < targetLength; i++)
+    {
+        text += possible.charAt(Math.floor(Math.random() * possible.length));
+    }
+    return text;
+};
